requirements.txt
black
isort
phidata
openai
pinecone-client
tiktoken
matplotlib
pandas
scikit-learn
tqdm
python-dotenv